requirements.txt
gradio
gradio[oauth]
fastapi
uvicorn[standard]
requests
langgraph
langchain_community
duckduckgo-search
tavily-python
langchain_tavily
langchain_google_genai
wikipedia
wikipedia-api
langchain_groq
pytesseract
numpy
pandas
matplotlib
pillow
openpyxl
arxiv
playwright
lxml
beautifulsoup4
youtube-transcript-api
faiss-cpu
sentence-transformers
huggingface-hub
pymupdf